pyproject.toml
[tool.poetry]
name = "ics-sync-server"
version = "0.1.0"
description = "Generate ICS files from other API to fill my personnal calendar automatically"
authors = ["Aurelien Labate <aurelien@labate.me>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
ics = "^0.7.2"
flask = "^3.1.0"
flask-caching = "^2.3.1"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
